Move Puck analytics to general Analytics file
